requirements.txt
pg8000
google-generativeai
python-telegram-bot
asyncio
python-dotenv
streamlit
openai
pandas
